+++ jackson/src/test/scala/com/twitter/finatra/jackson/tests/ScalaObjectMapperTest.scala
@@ -99,7 +99,7 @@ class ScalaObjectMapperTest extends AbstractScalaObjectMapperTest {
     result.fooClass should equal(FooClass("12345"))
   }
 
-  test("regular mapper handles unknown properties") {
+  test("regular mapper handles unknown properties while json provide MORE fields that case class") {
     // regular mapper -- doesn't fail
     mapper.parse[CaseClass](
       """
@@ -118,6 +118,30 @@ class ScalaObjectMapperTest extends AbstractScalaObjectMapperTest {
           |{
           |  "id": 12345,
           |  "name": "gadget",
+          |  "extra": "fail"
+          |}
+          |""".stripMargin
+      )
+    }
+  }
+
+  test("regular mapper handles unknown properties while json provide LESS fields that case class") {
+    // regular mapper -- doesn't fail
+    mapper.parse[CaseClassWithOption](
+      """
+        |{
+        |  "value": 12345,
+        |  "extra": "fail"
+        |}
+        |""".stripMargin
+    )
+
+    // mapper = loose, case class = annotated strict --> Fail
+    intercept[JsonMappingException] {
+      mapper.parse[StrictCaseClassWithOption](
+        """
+          |{
+          |  "value": 12345,
           |  "extra": "fail"
           |}
           |""".stripMargin